+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - [**BREAKING**] Passing empty array to `files` will now disable the config.
 - Set a new option `allowRethrowing: true` for [`@typescript-eslint/only-throw-error`](https://typescript-eslint.io/rules/only-throw-error) rule.
 - Enabled [`vue/no-custom-modifiers-on-v-model`](https://eslint.vuejs.org/rules/no-custom-modifiers-on-v-model.html) and [`vue/no-multiple-template-root`](https://eslint.vuejs.org/rules/no-multiple-template-root.html) rules for Vue 2 codebases.
+- Introduced a new root option `mode` with possible values of `app` and `lib`. It currently only controls whether you're allowed to import from `devDependencies` or not.
+- [**BREAKING**] Importing from `peerDependencies` is now allowed by default (rule: [`import/no-extraneous-dependencies`](https://github.com/un-ts/eslint-plugin-import-x/blob/HEAD/docs/rules/no-extraneous-dependencies.md)). Also disabled this rule in `cli` config.
 
 ### Dependencies